+++ parselglossy/api.py
@@ -27,17 +27,167 @@
 # parselglossy library, see: <http://parselglossy.readthedocs.io/>
 #
 
-# -*- coding: utf-8 -*-
 """Top-level functions for parselglossy."""
 
 import json
 from pathlib import Path
-from typing import Optional, Union
+from typing import List, Optional, Union
 
+from pyparsing import __file__ as ppfile
+
+from . import generation
 from .documentation import documentation_generator
+from .exceptions import ParselglossyError
 from .grammars import lexer
-from .utils import JSONDict, as_complex, path_resolver, read_yaml_file
+from .utils import JSONDict, as_complex, copier, path_resolver
 from .validation import is_template_valid, validate_from_dicts
+from .yaml_utils import read_yaml_file
+
+
+def generate(
+    template: Union[str, Path],
+    *,
+    where: Union[str, Path] = Path.cwd() / "input_parser",
+    grammar: Union[str, Path, List[Path]] = "standard",
+    tokenize: Optional[str] = None,
+    docfile: str = "input.rst",
+    doc_header: str = "Input parameters",
+) -> Path:
+    """Generate parser for client.
+
+    Parameters
+    ----------
+    template : Union[str, Path]
+        Which validation template to use.
+    where : Union[str, Path]
+        Where to generate the parsing module. Default to `input_parser` folder
+        under current working directory.
+    grammar : Union[str, Path, List[Path]]
+        The file containing the grammar to use to tokenize user input.
+        Defaults to `standard`, *i.e.* use one of the grammars packaged with
+        the library.
+    tokenize : Optional[str]
+        The commands to perform lexing of the input with a custom grammar. The
+        result of these commands must be a variable named `ir` of type
+        `Dict[str, Any]`.
+        Defaults to `None`.
+    docfile : str
+        The name of the documentation file for the input.
+        Defaults to `input.rst`.
+    doc_header : str
+        Header for the documentation page.
+        Defaults to ``Input parameters``.
+
+    Returns
+    -------
+    parser_dir : Path
+        Location of generated parser Python module
+
+    Notes
+    -----
+    This function will generate a Python module for parsing inputs as defined
+    by the template and grammar provided as parameters.
+
+    The user can provide a grammar as (a list of) external file(s).
+    There are a few constraints:
+       - The function performing tokenization **must return** an object of
+    ``Dict[str, Any]`` type: a, potentially recursive, dictionary of keys and
+    values.
+       - Commands to perform tokenization are passed *via* the `tokenize` input
+         parameters. This is a string that will copied *verbatim* in the
+         generated code, relevant `import` statements have to be included in
+         this string!
+       - The result of the commands in `tokenize` must be a variable named
+         `ir`.
+       - It is the responsibility of the user to satisfy any dependencies of
+         the parsing grammar they provided.
+
+    Raises
+    ------
+    :exc:`ParselglossyError`
+    """
+    where_ = path_resolver(where, touch=False)
+
+    # validate grammar options
+    if isinstance(grammar, (Path, list)):
+        if tokenize is None:
+            raise ParselglossyError(
+                "You need to set custom lexing commands when using a custom grammar."
+            )
+        else:
+            lexer_str = tokenize
+
+    # Create plumbing subfolder
+    (where_ / "plumbing").mkdir(parents=True, exist_ok=True)
+    # Generate __init__.py
+    with (where_ / "__init__.py").open("w") as f:
+        f.write(generation.INIT_PY)
+    # Generate README.md
+    with (where_ / "README.md").open("w") as f:
+        f.write(generation.README)
+    # Copy files
+    # b. copy exceptions.py, types.py, utils.py, validation.py,
+    # validation_plumbing.py, views.py file
+    for fname in [
+        "exceptions",
+        "types",
+        "utils",
+        "validation",
+        "validation_plumbing",
+        "views",
+    ]:
+        copier(Path(__file__).parent.absolute() / f"{fname}.py", where_ / "plumbing")
+    # c. Copy the grammar.
+    # By default this is our standard grammar, but could be one file provided
+    # by the user. The file is called "grammar.py"
+    if isinstance(grammar, str):
+        if grammar not in ["standard", "getkw"]:
+            raise ParselglossyError(f"Grammar {grammar} not available.")
+        for x in [
+            ppfile,
+            Path(__file__).parent.absolute() / "grammars/atoms.py",
+            Path(__file__).parent.absolute() / "grammars/getkw.py",
+        ]:
+            copier(x, where_ / "plumbing")
+        # extract `parse_string_to_dict` and write it into `getkw.py`
+        fun = generation.get_parse_string_to_dict()
+        with (where_ / "plumbing/getkw.py").open("a") as buf:
+            buf.write(fun)
+        lexer_str = f"""from . import getkw
+    lexer = getkw.grammar(has_complex={grammar == 'standard'})
+    ir = getkw.parse_string_to_dict(lexer, in_str)"""
+    elif isinstance(grammar, Path):
+        copier(grammar.resolve(), where_ / "plumbing/grammar.py")
+    elif isinstance(grammar, list):
+        for _ in grammar:
+            copier(_.resolve(), where_ / "plumbing")
+    else:
+        raise ParselglossyError("Type not recognized for grammar")
+
+    # d. generate lexer.py
+    with (where_ / "plumbing/lexer.py").open("w") as f:
+        f.write(generation.lexer_py(lexer_str))
+
+    # Generate api.py
+    template_ = path_resolver(template, touch=False)
+    stencil = read_yaml_file(template_)
+
+    # Check if the template is valid
+    is_template_valid(stencil)
+
+    with (where_ / "api.py").open("w") as f:
+        f.write(generation.api_py(stencil))
+    # Generate cli.py (uses argparse)
+    with (where_ / "cli.py").open("w") as f:
+        f.write(generation.CLI_PY)
+    # Generate documentation
+    (where_ / "docs").mkdir(parents=True, exist_ok=True)
+    docs = documentation_generator(stencil, header=doc_header)
+    outfile = path_resolver(where_ / f"docs/{docfile}")
+    with outfile.open("w") as o:
+        o.write(docs)
+
+    return where_
 
 
 def lex(