+++ tests/test_cli.py
@@ -34,6 +34,8 @@
 import json
 import re
 from pathlib import Path
+from shutil import rmtree
+import subprocess
 
 import pytest
 from click.testing import CliRunner
@@ -186,3 +188,80 @@ def test_cli_commands(command, out, reference):
 
     # Clean up file
     dumped.unlink()
+
+
+@pytest.mark.parametrize(
+    "command,inp,references",
+    [
+        (
+            [
+                "generate",
+                "--template",
+                "tests/validation/overall/template.yml",
+                "--target",
+                "fufamu",
+                "--grammar",
+                "standard",
+                "--docfile",
+                "foo.rst",
+            ],
+            "tests/api/scf.inp",
+            ("tests/ref/scf_fr.json", "tests/ref/generated_input.rst"),
+        ),
+        (
+            [
+                "generate",
+                "--template",
+                "tests/validation/overall/template.yml",
+                "--target",
+                "fufamu",
+                "-g",
+                f"{Path(__file__).parents[1].absolute() / 'parselglossy/grammars/atoms.py'}",
+                "-g",
+                f"{Path(__file__).parents[1].absolute() / 'parselglossy/grammars/getkw.py'}",
+                "--tokenize",
+                "from . import getkw; lexer = getkw.grammar(has_complex=True); ir = lexer.parseString(in_str).asDict()\n",
+                "--docfile",
+                "babar.rst",
+            ],
+            "tests/api/scf.inp",
+            ("tests/ref/scf_fr.json", "tests/ref/generated_input.rst"),
+        ),
+    ],
+    ids=["generate-default", "generated-custom-grammar"],
+)
+def test_cli_generate(command, inp, references):
+    runner = CliRunner()
+    result = runner.invoke(cli.cli, command)
+    assert result.exit_code == 0
+
+    parser_dir = Path.cwd() / f"{command[4]}"
+
+    # Check generated documentation agrees with reference
+    ref_rst = Path(references[1]).resolve()
+    dumped = parser_dir / f"docs/{command[-1]}"
+    with ref_rst.open("r") as ref, dumped.open("r") as o:
+        assert o.read() == ref.read()
+
+    # write a front-end using the generated module
+    front = Path("dwigt.py")
+    with front.open("w") as buf:
+        buf.write(f"from {command[4]} import cli\ncli.cli()")
+
+    # spawn a subprocess to run the generated parser's CLI and parse an input file
+    subprocess.run(["python", "dwigt.py", inp])
+
+    # Check final representation matches with reference
+    ref_json = Path(references[0]).resolve()
+    out = Path("scf_fr.json").resolve()
+    with ref_json.open("r") as ref, out.open("r") as o:
+        assert json.loads(o.read(), object_hook=as_complex) == json.loads(
+            ref.read(), object_hook=as_complex
+        )
+
+    # delete front-end and parsed output
+    front.unlink()
+    out.unlink()
+
+    # clean up generated parser
+    rmtree(parser_dir)