+++ tests/test_standard.py
@@ -118,7 +118,7 @@ str_array = [oops, repeated, "lorem", "IpSuM"]
 """
     with pytest.raises(ParselglossyError) as e:
         tokens = lexer.parse_string_to_dict(grammar, keywords)
-    assert 'A keyword is repeated. Please check your input.' in str(e.value)
+    assert "A keyword is repeated. Please check your input." in str(e.value)
     keywords = """
 int_array = [42]
 bool_array = [on, true, yes, False, True, false]
@@ -136,7 +136,7 @@ Something {
 """
     with pytest.raises(ParselglossyError) as e:
         tokens = lexer.parse_string_to_dict(grammar, keywords)
-    assert 'A keyword is repeated. Please check your input.' in str(e.value)
+    assert "A keyword is repeated. Please check your input." in str(e.value)
 
 
 def section(name):